+++ biostar/forum/templates/forms/form_inplace.html
@@ -8,15 +8,49 @@
     {#  Include the recaptacha key #}
     <input id="recaptcha_key" type="hidden" value="{{ captcha_key }}">
 
+
 {% endif %}
 
 <div class="ui basic form-wrap segment fit" id="inplace">
+    <script type="text/javascript" src="{% static 'pagedown/Markdown.Converter.js' %}"></script>
+    <script type="text/javascript" src="{% static 'pagedown-extra/pagedown/Markdown.Converter.js' %}"></script>
+    <script type="text/javascript" src="{% static 'pagedown/Markdown.Sanitizer.js' %}"></script>
+    <script type="text/javascript" src="{% static 'pagedown/Markdown.Editor.js' %}"></script>
+    <script type="text/javascript" src="{% static 'pagedown-extra/Markdown.Extra.js' %}"></script>
+    <script type="text/javascript" src="{% static 'pagedown_init.js' %}{% randparam %}"></script>
     <form class="ui compact form">
         {# Create a new comment #}
         {% if new %}
             <div class="field">
-                <div id="wmd-button-bar"></div>
-                <textarea id="wmd-input" class="wmd-input comment autocomplete" rows="9">{% include 'markdown_template.md' %}</textarea>
+                <div class="wmd-wrapper image-upload-enabled ">
+                    <div id="wmd-button-bar"></div>
+
+                    <textarea id="wmd-input" name="{{ form.content.name }}" class="wmd-input comment autocomplete"
+                              rows="9">{% include 'markdown_template.md' %}</textarea>
+                    <div class="pagedown-image-upload">
+                        <h2>Insert Image</h2>
+                        <div class="form-row">
+                            <div>
+                                <label class="label">From the web</label>
+                                <input class="url-input" type="text" placeholder="http://"/>
+                            </div>
+                        </div>
+                        <div class="form-row">
+                            <div>
+                                <label class="label">From your computer</label>
+                                <input class="file-input" type="file" name="image" id="file"
+                                       data-action="{% url 'pagedown-image-upload' %}" accept="image/*"/>
+                            </div>
+                        </div>
+                        <div class="submit-row">
+                            <input class="submit-input" type="submit" value="Save" name="_addanother">
+                            <p class="deletelink-box"><a href="#" class="close-image-upload deletelink">Cancel</a>
+                            </p>
+                        </div>
+                    </div>
+
+
+                </div>
             </div>
 
         {% else %}
@@ -35,8 +69,39 @@
             {# Edit the content #}
             <div class="field">
                 {% if post.is_toplevel %}{% include 'widgets/post_message.html' %}{% endif %}
-                <div id="wmd-button-bar"></div>
-                <textarea id="wmd-input" class="wmd-input autocomplete" rows="{{ rows }}">{{ post.content }}</textarea>
+
+                <div class="wmd-wrapper image-upload-enabled ">
+                    <div id="wmd-button-bar"></div>
+                    <textarea id="wmd-input" class="wmd-input autocomplete"
+                              rows="{{ rows }}">{{ post.content }}</textarea>
+
+
+                    <div class="pagedown-image-upload">
+                        <h2>Insert Image</h2>
+                        <div class="form-row">
+                            <div>
+                                <label class="label">From the web</label>
+                                <input class="url-input" type="text" placeholder="http://"/>
+                            </div>
+                        </div>
+                        <div class="form-row">
+                            <div>
+                                <label class="label">From your computer</label>
+                                <input class="file-input" type="file" name="image" id="file"
+                                       data-action="{% url 'pagedown-image-upload' %}" accept="image/*"/>
+                            </div>
+                        </div>
+                        <div class="submit-row">
+                            <input class="submit-input" type="submit" value="Save" name="_addanother">
+                            <p class="deletelink-box"><a href="#" class="close-image-upload deletelink">Cancel</a>
+                            </p>
+                        </div>
+                    </div>
+
+
+                </div>
+
+
             </div>
 
             {# Edit the tags #}
@@ -59,7 +124,7 @@
             accepted.</p>
 
         {# Preview segment #}
-        <div class="ui preview segment">{{ html |safe}}</div>
+        <div class="ui preview segment">{{ html |safe }}</div>
 
         {# Submit and cancel buttons #}
         <div class="field">
@@ -77,33 +142,30 @@
 
 
 <script>
-$(document).ready(function () {
-    var converter = new Markdown.getSanitizingConverter();
-    var editor = new Markdown.Editor(converter);
-    editor.run();
+    $(document).ready(function () {
 
-    $('.ui.dropdown').dropdown();
-    drag_and_drop();
-    $('form .preview').each(function () {
-        var text = $(this).closest('form').find('.wmd-input').val();
-        var form = $(this).closest('form');
-        highligh_preview(form, text);
-    });
-    // initialize autocomplete
-    var users = "{{ users_str }}".split(',');
-    autocomplete_users(users);
-    // initialize tags dropdown.
-    tags_dropdown();
+        $('.ui.dropdown').dropdown();
+        drag_and_drop();
+        $('form .preview').each(function () {
+            var text = $(this).closest('form').find('.wmd-input').val();
+            var form = $(this).closest('form');
+            highligh_preview(form, text);
+        });
+        // initialize autocomplete
+        var users = "{{ users_str }}".split(',');
+        autocomplete_users(users);
+        // initialize tags dropdown.
+        tags_dropdown();
 
-    $(this).on('click', '#inplace .cancel', function () {
-        cancel_inplace()
-    });
-     $(this).on('keyup', 'body', function (event) {
-         if(event.keyCode === 27) {
-             cancel_inplace()
-         }
-     });
+        $(this).on('click', '#inplace .cancel', function () {
+            cancel_inplace()
+        });
+        $(this).on('keyup', 'body', function (event) {
+            if (event.keyCode === 27) {
+                cancel_inplace()
+            }
+        });
 
-})
+    })
 </script>
