+++ biostar/forum/templates/widgets/post_body.html
@@ -1,5 +1,6 @@
 {% load forum_tags %}
 
+
 {#  Post anchor. #}
 <a name="{{ post.uid }}"></a>
