+++ biostar/recipes/static/engine.css
@@ -145,11 +145,6 @@ pre.toml_code {
     background-color: transparent !important;
 }
 
-.pagedown-image-upload{
-    transform: none !important;
-    left: 35% !important;
-    top: 33% !important;
-}
 
 .wmd-preview{
     padding:14px !important;
