+++ biostar/recipes/templates/widgets/job_summary.md
@@ -1,7 +1,9 @@
 {% load engine_tags %}
 
+{% spaceless %}
+{% if data.items %}
 Parameters used during the run:
-
-{% if data.items %}{% for key, obj in data.items %}{% if obj.label %} - {{ obj.label }}:{% if obj.name %}__{{ obj.name }}__{% else %}__{{ obj.value }}__{% endif %}
-{% endif %}{% endfor %}
-{% else %}No user parameters were set.{% endif %}
+{% for key, obj in data.items %}{% if obj.label %}- {{ obj.label }}:{% if obj.name %}__{{ obj.name }}__{% else %}__{{ obj.value }}__{% endif %}{% endif %}   
+{% endfor %}
+{% endif %}
+{% endspaceless %}
