+++ conf/ansible/server-deploy.yml
@@ -5,14 +5,6 @@
   tasks:
     - include_vars: variables.yml
 
-    - name: Install python requirements
-      shell: "source {{conda_dir}}/etc/profile.d/conda.sh &&
-      conda activate {{conda_env}} &&
-      pip install -r conf/requirements.txt"
-      args:
-        executable: /bin/bash
-        chdir: "{{ engine_dir }}"
-
     - name: Create data backup
       shell: "./server-backup.sh"
       args:
@@ -24,6 +16,14 @@
         repo: "{{ engine_git }}"
         dest: "{{ engine_dir }}"
         update: yes
+
+    - name: Install python requirements
+      shell: "source {{conda_dir}}/etc/profile.d/conda.sh &&
+      conda activate {{conda_env}} &&
+      pip install -r conf/requirements.txt"
+      args:
+        executable: /bin/bash
+        chdir: "{{ engine_dir }}"
 
     - name: Migrate the server
       shell: "./server-migrate.sh"
